feat(SlidingPiece): SlidingPiece 추상 클래스 생성
