+++ src/main/java/chess/domain/piece/Bishop.java
@@ -3,25 +3,11 @@ package chess.domain.piece;
 import chess.domain.color.Color;
 import java.util.Set;
 
-public class Bishop extends Piece {
-    private static Set<Direction> directions = Direction.getDiagonalDirection();
+public class Bishop extends SlidingPiece {
+    private static Set<Direction> DIRECTIONS = Direction.getDiagonalDirection();
 
     public Bishop(Position position, Color color) {
-        super(position, color);
-    }
-
-    @Override
-    public Set<Position> findMovablePositions(Position destination) {
-        Direction direction = position.findDirectionTo(destination);
-        if (!directions.contains(direction)) {
-            throw new IllegalArgumentException("이동할 수 없습니다.");
-        }
-        return position.findCourses(direction, destination);
-    }
-
-    @Override
-    public Bishop update(Position destination) {
-        return new Bishop(destination, color);
+        super(position, color, DIRECTIONS);
     }
 
     @Override
@@ -30,5 +16,10 @@ public class Bishop extends Piece {
             return PieceType.WHITE_BISHOP;
         }
         return PieceType.BLACK_BISHOP;
+    }
+
+    @Override
+    public Bishop update(Position destination) {
+        return new Bishop(destination, color);
     }
 }
