+++ src/main/java/chess/domain/piece/Rook.java
@@ -3,25 +3,11 @@ package chess.domain.piece;
 import chess.domain.color.Color;
 import java.util.Set;
 
-public class Rook extends Piece {
+public class Rook extends SlidingPiece {
     private static Set<Direction> directions = Direction.getFourDirection();
 
     public Rook(Position position, Color color) {
-        super(position, color);
-    }
-
-    @Override
-    public Set<Position> findMovablePositions(Position destination) {
-        Direction direction = position.findDirectionTo(destination);
-        if (!directions.contains(direction)) {
-            throw new IllegalArgumentException("이동할 수 없습니다.");
-        }
-        return position.findCourses(direction, destination);
-    }
-
-    @Override
-    public Rook update(Position destination) {
-        return new Rook(destination, color);
+        super(position, color, directions);
     }
 
     @Override
@@ -30,5 +16,10 @@ public class Rook extends Piece {
             return PieceType.WHITE_ROOK;
         }
         return PieceType.BLACK_ROOK;
+    }
+
+    @Override
+    public Rook update(Position destination) {
+        return new Rook(destination, color);
     }
 }
